три положения бокового меню

+++ resources/views/admin/pages/edit.blade.php
@@ -9,8 +9,8 @@
             @section('panel-page')
                 <div class="btn-group panel-page">
                     <a href="{{ URL::previous() }}" class="btn btn-outline-primary"><i class="fa fa-step-backward"></i></a>
-                    <button @click="save" class="btn btn-outline-success"><i class="fa fa-check"></i> СОХРАНИТЬ</button>
-                    <button @click="del" class="btn btn-outline-danger"><i class="fa fa-trash"></i> УДАЛИТЬ</button>
+                    <button @click.stop="save" class="btn btn-outline-success"><i class="fa fa-check"></i> СОХРАНИТЬ</button>
+                    <button @click.stop="destroy" class="btn btn-outline-danger"><i class="fa fa-trash"></i> УДАЛИТЬ</button>
                 </div>
             @endsection
             <ul class="nav nav-tabs">
@@ -57,33 +57,34 @@
             $.post("{{ route('pages.update', $page->id) }}",
                     {_method: 'PUT', _token: $('input[name="_token"]').val(), content: editor.getValue()},
                     function (data) {
-                        snackbar(data);
+                        snackbar(data.message);
                     }
             );
         }, readOnly: true
     });
 </script>
 <script>
-    window.Laravel = {csrfToken: '{{ csrf_token() }}'};
+window.Laravel = {csrfToken: '{{ csrf_token() }}'};
 
-
-    new Vue({
-        el: '.panel-page',
-        methods: {
-            save: function () {
-                var formData = {};
-                formData.content = editor.getSession().getValue();
-                this.$http.patch('{{ route('pages.update', $page->id) }}', formData).then((response) => {
-                snackbar(response.body);
+new Vue({
+    el: '.panel-page',
+    methods: {
+        save: function () {
+            var formData = {};
+            formData.content = editor.getSession().getValue();
+            this.$http.patch('{{ route('pages.update', $page->id) }}', formData).then((response) => {
+                snackbar(response.json().message);
             });
-            },
-            del: function () {
+        },
+        destroy: function () {
+            if(confirm("Вы действительно хотите удалить страницу?")) {
                 this.$http.delete('{{ route('pages.destroy', $page->id) }}').then((response) => {
-                    window.location.href= response.json().redirect_url;
+                    window.location.href = response.json().redirect_url;
                     snackbar(response.json().message);
-                });
-            }
+                })
+            } return false
         }
-    });
+    }
+});
 </script>
 @endpush
